backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
django-allauth>=0.44,<1.0
django-cors-headers>=3.10,<4.0
wagtail>=3.0,<5.0
django-modelcluster>=6.0,<7.0
django-taggit>=3.0,<4.0
psycopg2-binary>=2.9,<3.0
django-redis>=5.0,<6.0
Pillow>=8.0,<10.0
